# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Scrabble game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/search:
    get:
      operationId: searchUsers
      tags: [users]
      summary: Search users by username
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserPublic"
  /users/me:
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/me/push-token:
    post:
      operationId: registerPushToken
      tags: [users]
      summary: Register Expo push token
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushTokenInput"
      responses:
        "200":
          description: Token registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /users/{userId}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get a user's public profile
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublic"
        "404":
          description: User not found

  # Friends
...
  /stats:
    get:
      operationId: getMyStats
      tags: [stats]
      summary: Get current user stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerStats"
  /stats/{userId}:
    get:
      operationId: getUserStats
      tags: [stats]
      summary: Get a user's stats
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerStats"

  # Notifications
  /notifications/settings:
    get:
      operationId: getNotificationSettings
      tags: [notifications]
      summary: Get notification preferences
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationSettings"
    patch:
      operationId: updateNotificationSettings
      tags: [notifications]
      summary: Update notification preferences
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationSettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationSettings"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    # Auth schemas
...
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        notificationTurn:
          type: boolean
        notificationChat:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - email
        - avatarUrl
        - notificationTurn
        - notificationChat
        - createdAt
    UserPublic:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        avatarUrl:
          type: ["string", "null"]
      required:
        - id
        - username
        - avatarUrl
    ProfileUpdate:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 20
    PushTokenInput:
      type: object
      properties:
        token:
          type: string
      required:
        - token

    # Friend schemas
...
    PlayerStats:
      type: object
      properties:
        userId:
          type: string
        username:
          type: string
        gamesPlayed:
          type: integer
        gamesWon:
          type: integer
        gamesLost:
          type: integer
        winRate:
          type: number
        averageScore:
          type: number
        bestScore:
          type: integer
        bestWord:
          type: ["string", "null"]
        bestWordScore:
          type: integer
        totalPoints:
          type: integer
      required:
        - userId
        - username
        - gamesPlayed
        - gamesWon
        - gamesLost
        - winRate
        - averageScore
        - bestScore
        - bestWord
        - bestWordScore
        - totalPoints

    # Notification schemas
    NotificationSettings:
      type: object
      properties:
        notificationTurn:
          type: boolean
        notificationChat:
          type: boolean
      required:
        - notificationTurn
        - notificationChat
    NotificationSettingsUpdate:
      type: object
      properties:
        notificationTurn:
          type: boolean
        notificationChat:
          type: boolean